Revert "Resolve benchmark runs through spin asv"
This reverts commit 1f8d335ee44c0cb6b8c0bf11e0382ceddd0ce36a.
